requirements.txt
paceutils.egg==info
